[feat]: documentation + auth + product + users testing done

+++ server/src/controllers/carpool.controller.ts
@@ -7,6 +7,7 @@ import { Carpool, User } from '../types/entities';
 import { SearchType } from '../types/enums';
 import { AppError } from '../utils';
 import { UserService } from '../services/user.service';
+import { UpdateCarpoolInput } from '../schemas';
 
 // Get all carpools
 export const getAllCarpools = async (
@@ -58,6 +59,7 @@ export const getCarpoolById = async (
   }
 };
 
+// Create carpool
 export const createCarpool = async (
   req: Request<{}, {}, { payload: QueryDeepPartialEntity<Carpool> }>,
   res: Response,
@@ -89,6 +91,7 @@ export const createCarpool = async (
   }
 };
 
+// Update carpool
 export const updateCarpool = async (
   req: Request<
     { id: string },
@@ -100,6 +103,11 @@ export const updateCarpool = async (
 ) => {
   try {
     const { id } = req.params;
+    const user_id: string = res.locals.user.id;
+    const carpool = await findCarpoolById({ id });
+    if (!carpool) return next(new AppError(404, 'carpool not found'));
+    if (carpool.publisher_id !== user_id)
+      return next(new AppError(401, 'Unauthorized'));
     const { payload }: { payload: QueryDeepPartialEntity<Carpool> } = req.body;
     const carpoolService = new CarpoolService();
     const updatedCarpool = await carpoolService.updateResource(id, payload);
@@ -116,6 +124,7 @@ export const updateCarpool = async (
   }
 };
 
+// Delete carpool
 export const deleteCarpool = async (
   req: Request<{ id: string }>,
   res: Response,
@@ -123,6 +132,11 @@ export const deleteCarpool = async (
 ) => {
   try {
     const { id } = req.params;
+    const user_id: string = res.locals.user.id;
+    const carpool = await findCarpoolById({ id });
+    if (!carpool) return next(new AppError(404, 'carpool not found'));
+    if (carpool.publisher_id !== user_id)
+      return next(new AppError(401, 'Unauthorized'));
     const carpoolService = new CarpoolService();
     const deletedCarpool = await carpoolService.deleteResource(id);
     res.status(200).json({
@@ -138,19 +152,22 @@ export const deleteCarpool = async (
   }
 };
 
+// Join carpool
 export const joinCarpool = async (
-  req: Request,
+  req: Request<{ id: string }>,
   res: Response,
   next: NextFunction
 ) => {
   try {
     const user = res.locals.user;
     const uid: string = user.id;
-    const id: string = req.params.id;
+    const { id }: { id: string } = req.params;
+
+    if (!user) return next(new AppError(404, 'User not found'));
 
     const carpool = await findCarpoolById({ id });
     if (!carpool) return next(new AppError(404, 'Carpool not found'));
-    if (!user) return next(new AppError(404, 'User not found'));
+
     if (carpool.rider_count < carpool.capacity) carpool.rider_count++;
     else return next(new AppError(400, 'Carpool maximum capacity reached'));
 
@@ -164,6 +181,11 @@ export const joinCarpool = async (
         new AppError(404, 'Unauthorized request: Attempt to breach.')
       );
 
+    if (bookingUser.id === carpool.publisher_id)
+      return next(
+        new AppError(403, 'Forbidden. You cannot book your own carpool.')
+      );
+
     carpool.user_id.push(bookingUser);
 
     await carpool.save();
@@ -171,6 +193,44 @@ export const joinCarpool = async (
       status: 'success',
       data: { createBooking, bookingUser, carpool },
       message: 'Booking created successfully',
+    });
+  } catch (err: any) {
+    console.log('Error: (carpool.controller -> createBooking)', err);
+    if (err instanceof Error)
+      return next(new AppError(res.statusCode, err.message));
+    else return next(new AppError(400, 'Something went Wrong'));
+  }
+};
+
+export const leaveCarpool = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  try {
+    const user = res.locals.user;
+    const uid: string = user.id;
+    const id: string = req.params.id;
+
+    if (!user) return next(new AppError(404, 'User not found'));
+
+    const carpool = await findCarpoolById({ id });
+    if (!carpool) return next(new AppError(404, 'Carpool not found'));
+
+    if (uid === carpool.publisher_id)
+      return next(
+        new AppError(403, 'Forbidden. You cannot leave your own carpool.')
+      );
+
+    if (carpool.user_id.some(user => user.id === uid)) {
+      carpool.user_id = carpool.user_id.filter(user => user.id !== uid);
+    } else return next(new AppError(404, 'Something went wrong.'));
+
+    await carpool.save();
+    res.status(201).json({
+      status: 'success',
+      data: { carpool },
+      message: 'Carpool left successfully',
     });
   } catch (err: any) {
     console.log('Error: (carpool.controller -> createBooking)', err);
